Scroll tabs using mouse wheel
Not ideal, because it doesn't work when the mouse hovers over empty
areas. If we connect that signal to the notebook itself, it will also
trigger when scrolling the web page -- if the web_view doesn't handle
the scroll. The event then propagates upwards in the hierarchy and ends
up in the notebook.

Means we would have to wrap the web_view in a dummy evbox that catches
all scroll events. Nah.