Update main window title again
This is important for password managers like keepass.

Unlike the previous version of this code, we no longer traverse widget
hierarchies. Instead, a pointer to the real tab label widget is stored
on whatever widget happens to be returned by gtk_notebook_get_tab_label().
Luckily, there is an easy way to do this via g_object.